Tidied up code in new class and started adding temp card
A new class was made that contained the same code, but neater and with less spaghetti. There was an attempt made so that clicking on an attack card would deal damage to the enemyShip, however finding a way to actually access the mana and HP of the ship is proving to be the only issue here so far.