Battle Graphics Attempt 2
Background, two ships, and a mana + health meter is displayed. Updating the clock, thus updating mana via the procedure in the BattleMode class doesn't work, so it's temporary done via increasing it with another procedure. 
Need to make the clock work normally, display the mana and health bar for the enemy, and then find a way to damage either one of the ships.