Enemy Attack implemented.
The enemy attacks once its mana is full, dealing 5 damage to the player. This is sort of using the card classes, so implementing this with a player should be a bit easier (until graphic elements are involved.)
Game crashes when the player HP goes to 0. Have to figure out a way to implement death soon.