Added functionality to cards (Pressable) for play attack and bug fixed.
The bug where the game would crash when someone would die has been fixed, and the game simply stopped counting up with the timer.

On top of this, the player is now able to attack the enemy ship and defeat it by using their mana. This is done by clicking on the Att(ack) card do deal 4 damage by spending 5 mana.
The defense/armour variable has been added and works when coded in, but there's issues actually implementing it into the card click. 
Probably a simple mistake (Rounds everything but 100% armour down to 0), which I'll fix next time.