Added card draw + hand system.
This one was a pain.
Basically, this assumes that a deck has already been inputted from the travel mode, so instead I manually add some card data. JSON Parsing for it needs to be fixed too, eventually. May also need to pass through an array to show which CardIDs the player owns.
Anyway, the player's hand is randomly filled with cards from the deck. They use the card, it's put in the discard pile, and is replaced with another card from the deck, if there are any cards in the deck. If not, it's left empty. Once the deck and playerhand are empty, the discard pile is added back onto the deck, and all cards are drawn again.
I also deleted the old files, as they'll no longer be useful to me.
I'll create a to-do list of what needs to be done after all this.