Ship HP affects their sprite and 2 card sprite states, active or empty
The Ship sprites are basically done, adding an animation when they're attacking or repositioning it may be wanted though.
The cards only have 2 sprites, if there's a card there and if there's no card there. I believe that it'll be possible to make the card sprite vary by getting the CardID of each card and having the corresponding card sprite in the assets file to call, but that'll be quite a bit of effort and require more balancing.