Made card sprite depend on card's name plus more documentation
The card sprite will now be the "card's name" + .png. There is a bug where if there are two of the same cards, if one is pressed, the card in the other slot may be used up and placed in the discard pile as well. Not sure what causes it, possible boundary being set for the texture?
Documentation added, too, as well as a few new files to make it clear when cards are different.